feat: Merge all virtual storage nodes into a single node type.
This reduces some code duplication for what is mostly just
a different definition of the `reset` field.
