--- pywr-schema/src/nodes/monthly_virtual_storage.rs
@@ -1,204 +1,0 @@
-use crate::error::ComponentConversionError;
-#[cfg(feature = "core")]
-use crate::error::SchemaError;
-use crate::metric::{Metric, NodeComponentReference};
-#[cfg(feature = "core")]
-use crate::network::LoadArgs;
-use crate::node_attribute_subset_enum;
-#[cfg(feature = "core")]
-use crate::nodes::NodeAttribute;
-use crate::nodes::NodeMeta;
-use crate::nodes::core::StorageInitialVolume;
-use crate::parameters::Parameter;
-use crate::v1::{ConversionData, TryFromV1, try_convert_initial_storage, try_convert_node_attr};
-#[cfg(feature = "core")]
-use pywr_core::{
-    derived_metric::DerivedMetric,
-    metric::MetricF64,
-    virtual_storage::{VirtualStorageBuilder, VirtualStorageReset},
-};
-use pywr_schema_macros::PywrVisitAll;
-use pywr_v1_schema::nodes::MonthlyVirtualStorageNode as MonthlyVirtualStorageNodeV1;
-use schemars::JsonSchema;
-
-#[derive(serde::Deserialize, serde::Serialize, Clone, Debug, JsonSchema, PywrVisitAll)]
-#[serde(deny_unknown_fields)]
-pub struct NumberOfMonthsReset {
-    pub months: u8,
-}
-
-impl Default for NumberOfMonthsReset {
-    fn default() -> Self {
-        Self { months: 1 }
-    }
-}
-
-// This macro generates a subset enum for the `MonthlyVirtualStorageNode` attributes.
-// It allows for easy conversion between the enum and the `NodeAttribute` type.
-node_attribute_subset_enum! {
-    pub enum MonthlyVirtualStorageNodeAttribute {
-        Volume,
-        ProportionalVolume,
-    }
-}
-
-#[derive(serde::Deserialize, serde::Serialize, Clone, Default, Debug, JsonSchema, PywrVisitAll)]
-#[serde(deny_unknown_fields)]
-pub struct MonthlyVirtualStorageNode {
-    pub meta: NodeMeta,
-    /// Optional local parameters.
-    pub parameters: Option<Vec<Parameter>>,
-    pub nodes: Vec<NodeComponentReference>,
-    pub factors: Option<Vec<f64>>,
-    pub max_volume: Option<Metric>,
-    pub min_volume: Option<Metric>,
-    pub cost: Option<Metric>,
-    pub initial_volume: StorageInitialVolume,
-    pub reset: NumberOfMonthsReset,
-}
-
-impl MonthlyVirtualStorageNode {
-    const DEFAULT_ATTRIBUTE: MonthlyVirtualStorageNodeAttribute = MonthlyVirtualStorageNodeAttribute::Volume;
-
-    pub fn input_connectors(&self) -> Vec<(&str, Option<String>)> {
-        vec![]
-    }
-
-    pub fn output_connectors(&self) -> Vec<(&str, Option<String>)> {
-        vec![]
-    }
-
-    pub fn default_attribute(&self) -> MonthlyVirtualStorageNodeAttribute {
-        Self::DEFAULT_ATTRIBUTE
-    }
-}
-
-#[cfg(feature = "core")]
-impl MonthlyVirtualStorageNode {
-    /// This returns the node indices for flow constraints based on the nodes referenced in this virtual storage node.
-    ///
-    /// Note that this is a private function, as it is not supported using this node itself
-    /// inside a flow constraint.
-    fn node_indices_for_flow_constraints(
-        &self,
-        network: &pywr_core::network::Network,
-        args: &LoadArgs,
-    ) -> Result<Vec<pywr_core::node::NodeIndex>, SchemaError> {
-        let indices = self
-            .nodes
-            .iter()
-            .map(|node_ref| {
-                args.schema
-                    .get_node_by_name(&node_ref.name)
-                    .ok_or_else(|| SchemaError::NodeNotFound {
-                        name: node_ref.name.to_string(),
-                    })?
-                    .node_indices_for_flow_constraints(network, node_ref.component)
-            })
-            .collect::<Result<Vec<_>, _>>()?
-            .into_iter()
-            .flatten()
-            .collect();
-        Ok(indices)
-    }
-    pub fn add_to_model(&self, network: &mut pywr_core::network::Network, args: &LoadArgs) -> Result<(), SchemaError> {
-        let cost = match &self.cost {
-            Some(v) => v.load(network, args, Some(&self.meta.name))?.into(),
-            None => None,
-        };
-
-        let min_volume = match &self.min_volume {
-            Some(v) => Some(v.load(network, args, Some(&self.meta.name))?.try_into()?),
-            None => None,
-        };
-
-        let max_volume = match &self.max_volume {
-            Some(v) => Some(v.load(network, args, Some(&self.meta.name))?.try_into()?),
-            None => None,
-        };
-
-        let node_idxs = self.node_indices_for_flow_constraints(network, args)?;
-
-        let reset = VirtualStorageReset::NumberOfMonths {
-            months: self.reset.months as i32,
-        };
-
-        let mut builder = VirtualStorageBuilder::new(self.meta.name.as_str(), &node_idxs)
-            .initial_volume(self.initial_volume.into())
-            .min_volume(min_volume)
-            .max_volume(max_volume)
-            .reset(reset)
-            .cost(cost);
-
-        if let Some(factors) = &self.factors {
-            builder = builder.factors(factors);
-        }
-
-        network.add_virtual_storage_node(builder)?;
-        Ok(())
-    }
-
-    pub fn create_metric(
-        &self,
-        network: &mut pywr_core::network::Network,
-        attribute: Option<NodeAttribute>,
-    ) -> Result<MetricF64, SchemaError> {
-        // Use the default attribute if none is specified
-        let attr = match attribute {
-            Some(attr) => attr.try_into()?,
-            None => Self::DEFAULT_ATTRIBUTE,
-        };
-
-        let idx = network
-            .get_virtual_storage_node_index_by_name(self.meta.name.as_str(), None)
-            .ok_or_else(|| SchemaError::CoreNodeNotFound {
-                name: self.meta.name.clone(),
-                sub_name: None,
-            })?;
-
-        let metric = match attr {
-            MonthlyVirtualStorageNodeAttribute::Volume => MetricF64::VirtualStorageVolume(idx),
-            MonthlyVirtualStorageNodeAttribute::ProportionalVolume => {
-                let dm = DerivedMetric::VirtualStorageProportionalVolume(idx);
-                let derived_metric_idx = network.add_derived_metric(dm);
-                MetricF64::DerivedMetric(derived_metric_idx)
-            }
-        };
-
-        Ok(metric)
-    }
-}
-
-impl TryFromV1<MonthlyVirtualStorageNodeV1> for MonthlyVirtualStorageNode {
-    type Error = ComponentConversionError;
-
-    fn try_from_v1(
-        v1: MonthlyVirtualStorageNodeV1,
-        parent_node: Option<&str>,
-        conversion_data: &mut ConversionData,
-    ) -> Result<Self, Self::Error> {
-        let meta: NodeMeta = v1.meta.into();
-
-        let cost = try_convert_node_attr(&meta.name, "cost", v1.cost, parent_node, conversion_data)?;
-        let max_volume = try_convert_node_attr(&meta.name, "max_volume", v1.max_volume, parent_node, conversion_data)?;
-        let min_volume = try_convert_node_attr(&meta.name, "min_volume", v1.min_volume, parent_node, conversion_data)?;
-
-        let initial_volume =
-            try_convert_initial_storage(&meta.name, "initial_volume", v1.initial_volume, v1.initial_volume_pc)?;
-
-        let nodes = v1.nodes.into_iter().map(|n| n.into()).collect();
-
-        let n = Self {
-            meta,
-            parameters: None,
-            nodes,
-            factors: v1.factors,
-            max_volume,
-            min_volume,
-            cost,
-            initial_volume,
-            reset: NumberOfMonthsReset { months: v1.months },
-        };
-        Ok(n)
-    }
-}
