+++ pywr-schema/src/nodes/virtual_storage.rs
@@ -4,22 +4,24 @@ use crate::error::SchemaError;
 use crate::metric::{Metric, NodeComponentReference};
 #[cfg(feature = "core")]
 use crate::network::LoadArgs;
-use crate::node_attribute_subset_enum;
 #[cfg(feature = "core")]
 use crate::nodes::NodeAttribute;
 use crate::nodes::NodeMeta;
 use crate::nodes::core::StorageInitialVolume;
 use crate::parameters::Parameter;
 use crate::v1::{ConversionData, TryFromV1, try_convert_initial_storage, try_convert_node_attr};
+use crate::{ConversionError, node_attribute_subset_enum};
+use pywr_core::timestep::TimeDomain;
 #[cfg(feature = "core")]
-use pywr_core::{
-    derived_metric::DerivedMetric,
-    metric::MetricF64,
-    virtual_storage::{VirtualStorageBuilder, VirtualStorageReset},
-};
+use pywr_core::{derived_metric::DerivedMetric, metric::MetricF64, virtual_storage::VirtualStorageBuilder};
 use pywr_schema_macros::PywrVisitAll;
-use pywr_v1_schema::nodes::VirtualStorageNode as VirtualStorageNodeV1;
+use pywr_v1_schema::nodes::{
+    AnnualVirtualStorageNode as AnnualVirtualStorageNodeV1, MonthlyVirtualStorageNode as MonthlyVirtualStorageNodeV1,
+    RollingVirtualStorageNode as RollingVirtualStorageNodeV1, VirtualStorageNode as VirtualStorageNodeV1,
+};
 use schemars::JsonSchema;
+use std::num::NonZeroUsize;
+use strum_macros::{Display, EnumDiscriminants, EnumIter, EnumString, IntoStaticStr};
 
 // This macro generates a subset enum for the `VirtualStorageNode` attributes.
 // It allows for easy conversion between the enum and the `NodeAttribute` type.
@@ -27,6 +29,88 @@ node_attribute_subset_enum! {
     pub enum VirtualStorageNodeAttribute {
         Volume,
         ProportionalVolume,
+    }
+}
+
+#[derive(serde::Deserialize, serde::Serialize, Clone, Debug, JsonSchema, PywrVisitAll)]
+#[serde(deny_unknown_fields)]
+pub struct AnnualReset {
+    pub day: u8,
+    pub month: u8,
+    pub use_initial_volume: bool,
+}
+
+#[derive(serde::Deserialize, serde::Serialize, Clone, Debug, JsonSchema, PywrVisitAll, Display, EnumDiscriminants)]
+#[serde(tag = "type", deny_unknown_fields)]
+#[strum_discriminants(derive(Display, IntoStaticStr, EnumString, EnumIter))]
+#[strum_discriminants(name(VirtualStorageResetType))]
+pub enum VirtualStorageReset {
+    Never,
+    Annual(AnnualReset),
+    Monthly { months: u8 },
+}
+
+impl TryInto<pywr_core::virtual_storage::VirtualStorageReset> for VirtualStorageReset {
+    type Error = SchemaError;
+    fn try_into(self) -> Result<pywr_core::virtual_storage::VirtualStorageReset, Self::Error> {
+        let r = match self {
+            VirtualStorageReset::Never => pywr_core::virtual_storage::VirtualStorageReset::Never,
+            VirtualStorageReset::Annual(annual) => {
+                let reset_month = annual.month.try_into()?;
+                pywr_core::virtual_storage::VirtualStorageReset::DayOfYear {
+                    day: annual.day as u32,
+                    month: reset_month,
+                }
+            }
+            VirtualStorageReset::Monthly { months } => {
+                pywr_core::virtual_storage::VirtualStorageReset::NumberOfMonths { months: months.into() }
+            }
+        };
+
+        Ok(r)
+    }
+}
+
+/// The length of the rolling window.
+///
+/// This can be specified in either days or time-steps.
+#[derive(serde::Deserialize, serde::Serialize, Clone, Debug, JsonSchema, PywrVisitAll, Display, EnumDiscriminants)]
+#[serde(tag = "type", deny_unknown_fields)]
+#[strum_discriminants(derive(Display, IntoStaticStr, EnumString, EnumIter))]
+#[strum_discriminants(name(RollingWindowType))]
+pub enum RollingWindow {
+    Days { days: NonZeroUsize },
+    Timesteps { timesteps: NonZeroUsize },
+}
+
+impl Default for RollingWindow {
+    fn default() -> Self {
+        Self::Timesteps {
+            timesteps: NonZeroUsize::new(30).expect("30 is not zero"),
+        }
+    }
+}
+
+#[cfg(feature = "core")]
+impl RollingWindow {
+    /// Convert the rolling window to a number of time-steps.
+    ///
+    /// If the conversion fails (e.g. the number of days is less than the time-step duration) then `None` is returned.
+    pub fn as_timesteps(&self, time: &TimeDomain) -> Option<NonZeroUsize> {
+        match self {
+            Self::Days { days } => {
+                let ts_days = match time.step_duration().whole_days() {
+                    Some(d) => d as usize,
+                    // If the timestep duration is not a whole number of days then the rolling window cannot be specified in days.
+                    None => return None,
+                };
+
+                let timesteps = days.get() / ts_days;
+
+                NonZeroUsize::new(timesteps)
+            }
+            Self::Timesteps { timesteps } => Some(*timesteps),
+        }
     }
 }
 
@@ -56,6 +140,8 @@ pub struct VirtualStorageNode {
     pub min_volume: Option<Metric>,
     pub cost: Option<Metric>,
     pub initial_volume: StorageInitialVolume,
+    pub reset: Option<VirtualStorageReset>,
+    pub window: Option<RollingWindow>,
 }
 
 impl VirtualStorageNode {
@@ -120,15 +206,27 @@ impl VirtualStorageNode {
 
         let node_idxs = self.node_indices_for_flow_constraints(network, args)?;
 
-        // Standard virtual storage node never resets.
-        let reset = VirtualStorageReset::Never;
-
         let mut builder = VirtualStorageBuilder::new(self.meta.name.as_str(), &node_idxs)
             .initial_volume(self.initial_volume.into())
             .min_volume(min_volume)
             .max_volume(max_volume)
-            .reset(reset)
             .cost(cost);
+
+        // Standard virtual storage node never resets.
+        if let Some(r) = self.reset.clone() {
+            let reset = r.try_into()?;
+            builder = builder.reset(reset);
+        }
+
+        if let Some(window) = &self.window {
+            let rolling_window =
+                window
+                    .as_timesteps(args.domain.time())
+                    .ok_or_else(|| SchemaError::InvalidRollingWindow {
+                        name: self.meta.name.clone(),
+                    })?;
+            builder = builder.rolling_window(rolling_window);
+        }
 
         if let Some(factors) = &self.factors {
             builder = builder.factors(factors);
@@ -196,6 +294,151 @@ impl TryFromV1<VirtualStorageNodeV1> for VirtualStorageNode {
             min_volume,
             cost,
             initial_volume,
+            reset: None,
+            window: None,
+        };
+        Ok(n)
+    }
+}
+
+impl TryFromV1<AnnualVirtualStorageNodeV1> for VirtualStorageNode {
+    type Error = ComponentConversionError;
+
+    fn try_from_v1(
+        v1: AnnualVirtualStorageNodeV1,
+        parent_node: Option<&str>,
+        conversion_data: &mut ConversionData,
+    ) -> Result<Self, Self::Error> {
+        let meta: NodeMeta = v1.meta.into();
+
+        let cost = try_convert_node_attr(&meta.name, "cost", v1.cost, parent_node, conversion_data)?;
+        let max_volume = try_convert_node_attr(&meta.name, "max_volume", v1.max_volume, parent_node, conversion_data)?;
+        let min_volume = try_convert_node_attr(&meta.name, "min_volume", v1.min_volume, parent_node, conversion_data)?;
+
+        let initial_volume =
+            try_convert_initial_storage(&meta.name, "initial_volume", v1.initial_volume, v1.initial_volume_pc)?;
+
+        let nodes = v1.nodes.into_iter().map(|n| n.into()).collect();
+
+        let n = Self {
+            meta,
+            parameters: None,
+            nodes,
+            factors: v1.factors,
+            max_volume,
+            min_volume,
+            cost,
+            initial_volume,
+            reset: Some(VirtualStorageReset::Annual(AnnualReset {
+                day: v1.reset_day as u8,
+                month: v1.reset_month as u8,
+                use_initial_volume: v1.reset_to_initial_volume,
+            })),
+            window: None,
+        };
+        Ok(n)
+    }
+}
+
+impl TryFromV1<MonthlyVirtualStorageNodeV1> for VirtualStorageNode {
+    type Error = ComponentConversionError;
+
+    fn try_from_v1(
+        v1: MonthlyVirtualStorageNodeV1,
+        parent_node: Option<&str>,
+        conversion_data: &mut ConversionData,
+    ) -> Result<Self, Self::Error> {
+        let meta: NodeMeta = v1.meta.into();
+
+        let cost = try_convert_node_attr(&meta.name, "cost", v1.cost, parent_node, conversion_data)?;
+        let max_volume = try_convert_node_attr(&meta.name, "max_volume", v1.max_volume, parent_node, conversion_data)?;
+        let min_volume = try_convert_node_attr(&meta.name, "min_volume", v1.min_volume, parent_node, conversion_data)?;
+
+        let initial_volume =
+            try_convert_initial_storage(&meta.name, "initial_volume", v1.initial_volume, v1.initial_volume_pc)?;
+
+        let nodes = v1.nodes.into_iter().map(|n| n.into()).collect();
+
+        let n = Self {
+            meta,
+            parameters: None,
+            nodes,
+            factors: v1.factors,
+            max_volume,
+            min_volume,
+            cost,
+            initial_volume,
+            reset: Some(VirtualStorageReset::Monthly { months: v1.months }),
+            window: None,
+        };
+        Ok(n)
+    }
+}
+
+impl TryFromV1<RollingVirtualStorageNodeV1> for VirtualStorageNode {
+    type Error = ComponentConversionError;
+
+    fn try_from_v1(
+        v1: RollingVirtualStorageNodeV1,
+        parent_node: Option<&str>,
+        conversion_data: &mut ConversionData,
+    ) -> Result<Self, Self::Error> {
+        let meta: NodeMeta = v1.meta.into();
+
+        let cost = try_convert_node_attr(&meta.name, "cost", v1.cost, parent_node, conversion_data)?;
+        let max_volume = try_convert_node_attr(&meta.name, "max_volume", v1.max_volume, parent_node, conversion_data)?;
+        let min_volume = try_convert_node_attr(&meta.name, "min_volume", v1.min_volume, parent_node, conversion_data)?;
+
+        let initial_volume =
+            try_convert_initial_storage(&meta.name, "initial_volume", v1.initial_volume, v1.initial_volume_pc)?;
+
+        let window = if let Some(days) = v1.days {
+            if let Some(days) = NonZeroUsize::new(days as usize) {
+                RollingWindow::Days { days }
+            } else {
+                return Err(ComponentConversionError::Node {
+                    name: meta.name.clone(),
+                    attr: "window".to_string(),
+                    error: ConversionError::UnsupportedFeature {
+                        feature: "Rolling window with zero `days` is not supported".to_string(),
+                    },
+                });
+            }
+        } else if let Some(timesteps) = v1.timesteps {
+            if let Some(timesteps) = NonZeroUsize::new(timesteps as usize) {
+                RollingWindow::Timesteps { timesteps }
+            } else {
+                return Err(ComponentConversionError::Node {
+                    name: meta.name.clone(),
+                    attr: "window".to_string(),
+                    error: ConversionError::UnsupportedFeature {
+                        feature: "Rolling window with zero `timesteps` is not supported".to_string(),
+                    },
+                });
+            }
+        } else {
+            return Err(ComponentConversionError::Node {
+                name: meta.name.clone(),
+                attr: "window".to_string(),
+                error: ConversionError::MissingAttribute {
+                    attrs: vec!["days".to_string(), "timesteps".to_string()],
+                },
+            });
+        };
+
+        let nodes = v1.nodes.into_iter().map(|n| n.into()).collect();
+
+        let n = Self {
+            meta,
+            parameters: None,
+            nodes,
+            factors: v1.factors,
+            max_volume,
+            min_volume,
+            cost,
+            initial_volume,
+            reset: None,
+            window: Some(window),
         };
         Ok(n)
     }
